updated extractors
