+++ starlite/utils/__init__.py
@@ -1,5 +1,6 @@
 from .dependency import is_dependency_field, should_skip_dependency_validation
 from .exception import get_exception_handler
+from .extractors import ConnectionDataExtractor
 from .model import convert_dataclass_to_model, create_parsed_model_field
 from .predicates import is_async_callable, is_class_and_subclass, is_optional_union
 from .scope import get_serializer_from_scope
@@ -11,6 +12,7 @@ from .url import join_paths, normalize_path
 
 __all__ = [
     "AsyncCallable",
+    "ConnectionDataExtractor",
     "convert_dataclass_to_model",
     "create_parsed_model_field",
     "create_template_engine",
